Add quick start instructions to README
At current, if you land on this repository in GitHub it doesn't have much
information about what to do with the contents of the repo. This commit
updates the README to use standard formatting from our new Conjur repo template,
which adds certification level, requirements, and usage instructions sections.

It also adds in the step-by-step instructions from conjur.org, in case people
want to try the quick start from GitHub directly.

It is my hope that by adding this to GitHub, it will also make it easier to
provide instructions for updates to the quick start flow to the web team -
since we can point them to this README and ask that the website be updated
with the flow as written here.